docs(gitguard): restructure documentation and improve guides
Comprehensive update of GitGuard documentation:
• Add detailed templates customization guide
• Create new contributing guidelines
• Expand project philosophy and rationale
• Remove outdated draft documents
• Move legacy implementation notes to dedicated folder
