+++ PHILOSOPHY.md
@@ -27,20 +27,23 @@ As a builder and hacker at heart, I believe in creating tools that make developm
 - Consistent patterns across all packages
 
 ## Current Focus
-This monorepo contains my essential developer toolkit:
+This monorepo contains my essential developer toolkit, representing a subset of tools I've developed over the years:
 
 - **@siteed/design-system**: A cross-platform UI component library
 - **@siteed/react-native-logger**: Advanced logging for mobile apps
 - **@siteed/publisher**: Streamlined monorepo release management
 - **@siteed/gitguard**: AI-powered commit quality tools
 
+While these packages are actively used in my daily projects, they represent just the beginning. I have numerous other tools and libraries developed over my career that will be gradually integrated, refined, and open-sourced here.
+
 ## Goals
 
 ### Short-term
 1. Improve documentation and examples
-2. Add more test coverage
-3. Create video tutorials and blog posts
-4. Build a community of contributors
+2. Create video tutorials and blog posts
+3. Integrate existing projects into the monorepo structure
+4. Establish consistent quality standards across all packages
+5. Share development insights through siteed.net
 
 ### Long-term
 1. Expand to hardware/software integration tools
@@ -48,7 +51,19 @@ This monorepo contains my essential developer toolkit:
 3. Create comprehensive project templates
 4. Build a ecosystem of compatible tools
 
+## Building in Public
+I believe in continuous improvement and learning in public. While these packages are still in their early stages and actively being refined, they represent real solutions I'm developing for common challenges I face in my projects. By developing in public, I:
+
+- Hold myself accountable for maintaining high code quality
+- Create opportunities for community feedback and collaboration
+- Share my development journey and learning process
+- Document the evolution from prototype to production-ready tools
+
+As these packages mature and prove themselves in real-world applications, they may evolve into their own dedicated repositories. This approach allows for transparent iteration and improvement while working towards battle-tested, production-grade solutions.
+
+You can follow my work and insights at [siteed.net](https://siteed.net).
+
 ## Contributing
-This project is my way of giving back to the developer community. While these tools are primarily built for my own needs, I welcome contributions that align with the core principles above.
+While these tools are primarily built for my own needs, I welcome contributions that align with the core principles above.
 
 See [CONTRIBUTING.md](./CONTRIBUTING.md) for guidelines.
