--- packages/gitguard/IDEA.md
@@ -1,23 +1,0 @@
-# @siteed/gitguard
-
-A smart git assistant that improves commit quality using AI. It analyzes your changes, suggests meaningful commit messages, and helps maintain a clean git history in both standard and monorepo projects.
-
-## Core Features
-
-- 🤖 **Smart Git Wrapper**
-  - Intercepts git commits to suggest improvements
-  - Preserves existing git workflow
-  - Zero config needed for basic usage
-  - Automatic repository structure detection
-
-- 🔍 **Intelligent Change Analysis**
-  - Parses and categorizes git diffs
-  - Groups related changes
-  - Detects breaking changes
-  - Creates meaningful summaries
-
-- 📦 **Monorepo Support**
-  - Automatic package detection
-  - Smart commit splitting
-  - Dependency-aware commit ordering
-  - Scope suggestions
