+++ packages/gitguard/docs/draft_pr_review.feature.md
@@ -1,0 +1,5 @@
+  -> Triggered via github action or manually
+  -> New prompt to review PR
+  -> Chains Of thought review analysis
+    -> Analysis grouping of feature in the PR and review them one by one
+    -> Each analysis will be in a new commit
