+++ packages/gitguard/docs/why.md
@@ -1,11 +1,87 @@
-WIP:
+# Why GitGuard?
 
-- document explaining some of the key reasons for building gitguard
+## Personal Journey
 
-I have seen the way I code change incredibly in the last year, and it seems to keep increasing as AI models improve.
-I need to find a way to be the most efficient leveraging these models without breaking existing codes or the bank.
+As someone who has been working with AI models for the past 3 years, primarily focusing on audio and vision processing, I've witnessed a dramatic shift in how I approach software development. The arrival of AI-powered IDEs like Cursor has completely transformed my programming paradigm. I've evolved from being just a programmer to becoming more of an orchestrator - directing AI to help me build software more efficiently.
 
-New feature to maintain api consistency and integrity.
-The problem with new development with code being generated by AI is that it's not always consistent with the existing codebase.
-This feature will help to maintain the consistency and integrity of the codebase by suggesting changes to the commit message and split.
-We also need to be able to track breaking changes and other issues in the codebase.
+My workflow continues to evolve as I learn to better balance AI assistance with code quality. While I'm getting better at this balance, particularly as AI models improve, it's an ongoing process of refinement and learning. Each interaction with AI tools teaches me new ways to maintain high standards while leveraging automation.
+
+A significant part of my journey involves exploring fully offline solutions. I spend considerable personal time working on "light" models that can run efficiently on local GPU clusters. This involves a mix of fine-tuning existing models and training new ones, with the ultimate goal of creating a specialized GitGuard model that can run completely offline while maintaining high performance.
+
+## The Challenge
+
+GitGuard emerged from a simple need: improving code quality while leveraging AI's capabilities. I wanted a tool that would:
+
+1. **Integrate Seamlessly**: Work alongside existing tools like:
+   - AI-powered IDEs
+   - Git workflows
+   - CI/CD pipelines
+   - Existing project conventions
+
+2. **Universal Compatibility**: Work across different contexts:
+   - Personal projects
+   - Professional work environments
+   - Open source contributions
+   - Team collaborations
+   - Any existing Git workflow
+
+3. **Enforce Quality**: Automatically check for:
+   - Consistent commit messages
+   - Proper PR structure
+   - Breaking changes
+   - Code patterns
+
+4. **Enhance Productivity**: Help developers:
+   - Generate meaningful commit messages
+   - Structure PRs effectively
+   - Split large changes logically
+   - Maintain project standards
+
+5. **Support Offline Workflows**: Enable local AI processing:
+   - Run with local models
+   - Operate without internet connectivity
+   - Maintain data privacy
+   - Reduce operational costs
+
+## The Solution
+
+GitGuard emerged from a simple need: maintaining code quality while leveraging AI's capabilities. I wanted a tool that would:
+
+1. **Integrate Seamlessly**: Work alongside existing tools like:
+   - AI-powered IDEs
+   - Git workflows
+   - CI/CD pipelines
+
+2. **Enforce Quality**: Automatically check for:
+   - Consistent commit messages
+   - Proper PR structure
+   - Breaking changes
+   - Code patterns
+
+3. **Enhance Productivity**: Help developers:
+   - Generate meaningful commit messages
+   - Structure PRs effectively
+   - Split large changes logically
+   - Maintain project standards
+
+## Evolution
+
+GitGuard started as a simple Python git prepare-commit-msg hook that I wrote one morning. It was designed to check commit message formats and automatically add suggestions. The initial success and potential of this simple script motivated me to:
+
+1. Migrate to TypeScript for better maintainability
+2. Expand the feature set significantly
+3. Create a full-fledged project with proper architecture
+4. Add comprehensive AI integration capabilities
+
+> You can still find the initial [Python script](../legacy/gitguard-prepare.py) and its original [README](../legacy/README.md) in the legacy folder.
+
+## Looking Forward
+
+As AI continues to evolve and reshape how we write code, tools like GitGuard will become increasingly important in:
+- Maintaining consistency across AI-assisted development
+- Ensuring quality standards are met
+- Helping teams adapt to new development paradigms
+- Creating sustainable, maintainable codebases
+
+
+The ultimate vision is to provide developers with a powerful, AI-enhanced workflow tool that can run entirely on their own infrastructure while maintaining the high quality standards that modern software development demands.
