--- packages/publisher/ISSUE
@@ -1,27 +1,0 @@
-# Changelog
-
-All notable changes to this project will be documented in this file.
-
-The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/),
-and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
-
-## [Unreleased]
-
-- fix: invalid changelog format
-
-## [0.4.11] - 2024-10-31
-- ⚠️ **WARNING: DEVELOPMENT IN PROGRESS** ⚠️
-
-## [0.4.10] - 2024-10-30
-- dry run mode
-- feat: dry run mode
-
-
-## [0.4.9] - 2024-10-30
-- full process with helper scripts
-- feat: full process with helper scripts
-
-[unreleased]: https://github.com/deeeed/universe/compare/@siteed/publisher@0.4.11...HEAD
-[0.4.11]: https://github.com/deeeed/universe/compare/@siteed/publisher@0.4.10...@siteed/publisher@0.4.11
-[0.4.10]: https://github.com/deeeed/universe/compare/@siteed/publisher@0.4.9...@siteed/publisher@0.4.10
-[0.4.9]: https://github.com/deeeed/universe/compare/@siteed/publisher@@siteed/publisher@0.4.8...@siteed/publisher@@siteed/publisher@0.4.9
